% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/talkinturkey.R
\name{turkeytalk}
\alias{turkeytalk}
\title{turkeytalk}
\usage{
turkeytalk(phrase = "Happy Thanksgiving!")
}
\arguments{
\item{phrase}{A character string containing the phrase the turkey will say. Default is "Happy Thanksgiving!"}
}
\value{
A character string representing a turkey saying the specified phrase.
}
\description{
Create a turkey that can speak
}
\examples{
cat(turkeytalk("Gobble gobble!"))
}
